A couple of important fixes for the reconstruction. One is to fix #154, making sure the min & max scales are applied appropriately. Second is to add a new parameter reconConvergence, which takes the previously hard-coded convergence criterion and makes it available to be altered. Documentation is updated for this parameter as well.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@1026 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/ATrous/filter.hh
@@ -60,7 +60,7 @@ namespace duchamp
     void   define(int code);
 
     /// @brief Calculate the number of scales possible with a given filter and data size. 
-    unsigned int    getNumScales(long length);
+    unsigned int    getNumScales(size_t length);
 
     /// @brief Calculate the maximum number of pixels able to be analysed with a filter at a given scale. 
     unsigned int    getMaxSize(int scale);
